requirements.txt
streamlit
numpy 
pandas
seaborn
matplotlib
plotly
scikit-learn
opencv-python
Pillow